backend/ipp.c:     - compress_files(): Fix debug message to use CUPS_LLCAST.
cups/testhttp.c:
    - main(): Fix printf's to use CUPS_LLCAST on long long
      values.

cups/testppd.c:
    - main(): Fix printf's to use CUPS_LLCAST on long long
      values.

locale/translate.c:
    - translate_messages(): Cast the length of the request
      buffer.

scheduler/client.c:
    - pipe_command(): Cast the CONTENT_LENGTH value.

scheduler/dirsvc.c:
    - cupsdSendCUPSBrowse(): Cast the browser number in the
      error message.

scheduler/subscriptions.c:
    - cupsdAddEvent(): Use job->state_value, not the pointer
      value...


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@5261 7a7537e8-13f0-0310-91df-b6672ffda945

+++ cups/testppd.c
@@ -111,8 +111,8 @@ main(int  argc,				/* I - Number of command-line arguments */
     puts("PASS");
   else
   {
-    printf("FAIL (%d bytes instead of %d)\n", s ? strlen(s) : 0,
-           strlen(default_code));
+    printf("FAIL (%d bytes instead of %d)\n", s ? (int)strlen(s) : 0,
+           (int)strlen(default_code));
 
     if (s)
       puts(s);
